the disk root is built from DESTDIR, not from the ramdisk image
Until now mkroot.sh and "ramimage.sh -d" took the ramdisk's proto whole: the
root on the disk was the boot image on a medium that remembers what is
written to it.  That was right while the ramdisk was the only thing that
existed.  It is wrong now, and it cannot simply be grown: the ramdisk is
linked into the memory driver, rides in boot.mba and stays in RAM forever.

So the two are split.  evbarm64_rootproto.py parses the ramdisk's generated
proto into a tree, adds the userland directories from DESTDIR (bin, sbin,
usr/bin, usr/sbin, usr/libexec, usr/man, usr/share) and prints the merged
proto with absolute paths.  Where a name is in both, the ramdisk wins: its
/etc, its device nodes and its boot servers are what brings the system up,
and they are written down literally -- modes, owners, major and minor -- in
a way DESTDIR does not record.

usr/lib and usr/include are left out: everything links static, and headers
and library archives are of use only to a compiler, which the board has not
got.

The same code makes the QEMU disk and the eMMC root, so the difference
between them stays exactly one: the medium.  60 MB of programs, so the QEMU
image grows from 64 MB to 512.

Three things the new root needed and did not have:

  /etc/man.conf     man(1) reads it for the search path and for which
                    program renders which suffix.  Without it there are 210
                    manual pages on the disk and no way to reach them.
                    Manual pages live in /usr/man on MINIX, not
                    /usr/share/man, which is why they looked missing.

  /var/tmp/vi.recover  vi says so on every start otherwise, over half a
                    screen, and waits for a keypress.

  usr.bin/man       did not build. -- [code age]

And one bug found by running the result rather than by reading it: the libc
ptrace() wrapper was declared "int ptrace(int, pid_t, void *, int)", but
MINIX's ptrace is not NetBSD's -- T_GETINS, T_GETDATA and T_GETUSER return a
whole *word* of the traced process or of its kernel structures, and that
word travels in a message field declared long.  On ILP32 the two were the
same; on LP64 the top half of every read was thrown away.  Now long, under
defined(__minix) so NetBSD's own prototype is untouched. -- [aarch64 gap]

That is the second time in this port that a 32-bit field silently cut a
pointer-sized value in half (VFS_PM_PS_STR in 5.5 was the first).  It did
not, however, turn out to be the whole of why trace(1) reports "Kernel magic
check failed": it still does, with the wrapper provably fixed (the
disassembly loads the full 64 bits).  offsetof(struct proc, p_magic) and
sizeof(struct proc) are identical when compiled with trace's flags and with
the kernel's, so the cause is elsewhere and is not yet known.  Recorded
rather than guessed at.

+++ minix/lib/libc/sys/ptrace.c
@@ -5,7 +5,7 @@
 #include <string.h>
 #include <sys/ptrace.h>
 
-int ptrace(int req, pid_t pid, void *addr, int data)
+long ptrace(int req, pid_t pid, void *addr, long data)
 {
   message m;
